♻️ Splitt ut ta av vent-funksjonaliteten i egen komponent
